[Image Orientation] Pipe orientation through from worker-decoded images
The image orientation is not applied when creating a CanvasImageSource
directly from an uploaded file. This is because we do not extract and
pass on the orientation from the decoder.

Fix it by obtaining orientation from the decoder on the worker thread
and passing it back to the original thread to create the image.

Bug: 1069965
Change-Id: I3f64519ed8b91e97bfa9afc802fe90ceb7eb2c3e