[@property] Improve test coverage for valid/invalid rules
This adds new tests that verify that invalid rules are ignored,
and don't affect other valid rules.

I also rewrote this test file to use utils.js more liberally. It makes
individual tests harder to understand for the uninitiated, but it
scales so much better as the number of test functions increases.

Bug: 973830
Change-Id: Ie19d85042a0475180cc4bbee4ad202ea08adba73