Implement checking for "is declarative shadow root" in attachShadow()
With this CL, if a declarative shadow root is already present when
attachShadow() is called, the contents of this shadow root are
removed, and the existing shadow root is returned. This is per Step 5
of the new spec [1], to allow coexistence of declarative content and
custom element hydration code.

This CL also adds an exhaustive test of all element types, both those
supporting and disallowing shadow root attachment.

[1] https://whatpr.org/dom/858.html#concept-attach-a-shadow-root

Bug: 1042130
Change-Id: Ie9fc39eeb324934f3cb13ab2eb6bc8a82d99d667